feat: order details

+++ app/src/main/java/com/selsela/takeefapp/data/order/model/order/Service.kt
@@ -13,5 +13,5 @@ data class Service(
     @SerializedName("name")
     val name: String = "",
     @SerializedName("price")
-    val price: Int = 0
+    val price: Double = 0.0
 )
